Fall back to user scheme
It appears to be default xcode behaviour to locate
the scheme under a user path and not a shared path.
To ensure that this step will work with default
xcode projects this change will search under the
project configuration directory for a user scheme
matching the scheme name configured in the workflow.